Split reusable report page content into partials
Page footer text is shared between pages so it makes
sense to have it in one place and include it as
needed.
Note: for now we're not going to show the footer

On preparing and ready page we play back what type
of notifications with what status user has
requested the report for. It makes to make it as
a macro and call it with data as needed.