Simple polling script to check report status
When the user in on the 'preparing report' page
this script periodicly checks a JSON endpoint
if the reports status has been updated

If the returned status does not not match expected
state, it keeps checking.

Once the status matches either a success or failed
state, teh script:

- updates the page content with wording about the
update state and notice of redirection

- waits for 10s (to give screenreaders time to
annouce the updated text and users to read the text)
and the reloads the page and the applciation handles
the routign to teh correct page.

We use location.replace https://developer.mozilla.org/en-US/docs/Web/API/Location/replace
so that the the current page is not stored in the
history, meaning the user won't be able to use the
back button to navigate to it. This way we hopefully
avoid the back button confusion, as the application
shows you the correct page. Back button, once JS
executes will take you back to the notifications
page.

Additional refactoring by Tom B.