Brings in the following: - https://github.com/alphagov/govuk-frontend/releases/tag/v5.9.0 - https://github.com/alphagov/govuk-frontend/releases/tag/v5.8.0 - https://github.com/alphagov/govuk-frontend/releases/tag/v5.7.1
None of those changes effect this app. We set the
`$govuk-new-organisation-colours` flag here just
to silence the deprecation warnings you see when
compiling your Sass.

Note that this also brings in v3.5.0 of the GOVUK
Frontend Jinja Macros for our HTML rendering,
which is the version for v5.9.0 of GOVUK Frontend:

https://github.com/LandRegistry/govuk-frontend-jinja?tab=readme-ov-file#compatibility