Add daily limit for international sms for new services
After much discussion, we decided to set this limit at 100.

Factors that went into that decision:
- we wanted the limit low enough to mitigate against AIT attacks,
even if they are "trickle" attacks.
- we wanted the limit high enough to not interfere with most services

100 sms a day was a good compromise - it will lead to around £20
maximum daily bill for international sms, and at the same time
it is above what most services are sending.

From queries we ran, we estimate that each year around 5 services
will run into this limit - even if they do need our support,
that is a manageable number.

Hopefully they will find the information where we tell them they
can change their limit.

Yet we are afraid they might not find that information, or even
not notice they are hitting the limit and silently drop notifications.

Because of this, we will have a think about monitoring when someone
hits the new international_sms_message_limit.