Make test for add_service more useful
Recently, a deploy broke on adding a service, because we forgot
to add the new argument we are setting to create_service
function on ServiceApiClient.

This was not caught by our unit tests, because we mock too early
- we mock the whole client, instead of just mocking the post
request it makes.

I changed the test to do mocking later, at the point when we call
the API. This way a similar issue will be caught in the future.