Update main templates with to support rebrand
Until the rebrand lauch date we still need to serve
current static assets, header and footer styles
and then be able to quickly update on the day.

To support this, the template has been tweaked with
a `govukRebrand` Boolean which if `true`
will turn on new header and footer styles and well
as update minor favicon colors.

Our custom footer and the govukHeader have
been set up with a 'rebrand' option which is
controlled by the `govukRebrand` setting.

To switch to the rebrand assets, we need to set
`govukRebrand: true`, merge and deploy.