Change go-live URLs and add redirects for old URLs
This changes the request to go live URLs from
`/services/<uuid:service_id>/service-settings/request-to-go-live` to
`/services/<uuid:service_id>/make-your-service-live` and
adds in redirects for the old routes.

I've kept the view function names the same, since they still work for
the new URLs and this reduces the amount of code we need to change.