Show conditional content on 'make your service live' page
When a request is submitted, the user is taken
back to the page and shown info about submission
instead of seeing a completed form again.
We show the same content about the person who
submitted the form if a service has an active
go-live request and page is viewed by another
team member.

If the submitter is the same person as approver,
we say "you", otherwise we provide the name of
the go-live requester.