Use mock to tell tests we’re in business hours
Using `freeze_time` causes problems because it only freezes the time
during the test run, not during fixture setup.

This can make the session invalid, which means the code can’t get stuff
like the current service from the session.

The reason we are freezing time is to make sure these tests don’t go
through the out-of-hours support route. So let’s do that explicitly with
a mock instead.