Remove uses of `datetime.utcnow`
`datetime.utcnow` is deprecated as of Python 3.12:
https://docs.python.org/3/library/datetime.html#datetime.datetime.utcnow

The docs say:
> Use `datetime.now()` with `datetime.UTC` instead.

There are a few cases where we pass serialised datetimes to the API.
Because the API might be expecting datetimes in a particular format I’ve
written a couple of helper methods to deal with this. In the future we
could investiagate whether the API really needs datetimes in those
specific formats.