Change left hand nav text for self-service org approvers
From "Make service live" to "Make this service live". This is done to
help differentiate it from the new menu item that will be added, "Make
your service live".

This also ensures that the menu item is higlighted when on the page to
make the go-live decision, which was missing from the list.