Fix border-bottom on header
Fixes for 2 problems:
- the transparent border-bottom on the header
  component showing up as white on the product
  page
- the white border-top we set on the intro section
  doubling the border you see on mobile

Design system put a transparent border-bottom on
the header component to be invisible to most users
but visible in high contrast modes to mark its
boundaries. On our product page, this shows up
because the top section is made up of blue
sections on a white background. The transparent
border therefore creates a gap through which you
see the white background.

We already have a white-ish border below the
header on the product page so this also looks like
we've got 2 of them when the header's
border-bottom shows.

This sets the header's border-bottom to the same
white-ish colour we already set on the intro
section border-top. We also swap between these
borders at different screen sizes, showing the
header's border-bottom on smaller screens and the
intro section's border-top on any above that.