Fix ‘today at midnight’ being labelled ‘tomorrow at midnight’ during British Summer Time
`formatters.get_human_day` was expecting a UTC datetime but was being
given a localized one.

This meant that during British Summer Time it was an hour out.

Our tests only covered winter, so that’s why we didn’t spot this when
we retired the `forms.get_human_day` function, which silently expected
a localized datetime.