Stop rebranded header navigation from jumping around
Because the header navigation now uses bold for its active state each
item changes in width when clicked.

This causes the other items to shift position which looks a bit janky.

We can use a similar trick that we do in the side navigation to make
things an almost-consistent width:
https://github.com/alphagov/notifications-admin/blob/2456d714fc54f45cbd981e1717ad0533dcb5241b/app/assets/stylesheets/components/navigation.scss#L120-L123