Allow platform admin users to see the sending journey
Platform admin users should not be allowed to send messages on behalf of
a service (either maliciously or by accident).

To prevent this we restrict any endpoints related to sending using the
`restrict_admin_use` flag. However this stops us from seeing these pages
as our users do, which makes it harder to debug issues when they occur.

This commit makes the steps of the journey visible to platform admin
users, while still leaving a restriction on the final step (either
sending a job or sending a one-off message).