Don’t serve session cookie from the no_cookie endpoint
Flask has called blueprints which are a way of grouping together
endpoints with similar functions.
In the admin app we have 3:
- `main` – serves most pages, eg /dashboard, /templates, /send-one-off,
  etc.
- `json_updates` – serves the .json responses which make the dashboard
  etc. update periodically
- `no_cookie` – serves images of letters and the preview of email
  branding

For every request, the browser sends a session cookie and the app
responds with an updated session cookie. We store the personalisation
for sending letters (for example address, name, reference) in this
cookie. This change tells the server to only send an updated cookie from
the main blueprint. This means a stale request from the `no_cookie`
endpoint won’t override a more recent request. I think this is the way
we thought it was working, but it wasn’t.