Increase button spacing on /uploads page
On mobile viewport buttons to "upload a ltter" and "upload an emergecy contact list" have no space between them.

This adds a Design System 15px spacing class to the first button.

While this does slightly increase the hight of the container when the second button is not present, this is negligable in the wider context of that container