Increase button spacing on /email-branding page
On mobile viewport button to "add email branding" and "add government identity logo" have no space between them.

This adds a Design System 15px spacing class to the first button.

While this does slighlty increase the hight of the container when the second button is not present, this is negligable in the wider context of that container