Fix the platform admin search for service callbacks
This code expected the `/platform-admin/find-by-uuid` api endpoint
to return a different model name based on the type of service callback
being searched for. However, since all types of service callback are in
the `service_callback_api` table, it always returns that regardless
of what type of callback id was searched for.

This updates the code to check for what the api actually returns. Making
the search functionality take you to the actual type of callback would
require a bit of a redesign, but as a simple fix you now get taken to
the page listing all the callbacks for the service.