Rewrite LiveCheckboxControls as ESM module
We want to migrate all of our legacy Javascript
files to ESM, so this is the latest in the series
of conversions.

Rewrite follows the same conventions we previously
used for:
- FocusBanner
- Collapsible checkboxes
- ColourPreview
- FileUpload
- Autofocus
- PreviewPane
- CopyToClipboard
- LiveSearch
- ListEntry
- RadiosWithImages
- EnhancedTextbox

LiveCheckboxControls is a base class for both
AddBrandingOptionsControls and SetAuthTypeForm so
I did them all together.

This also includes a rewrite of their tests, not
just to allow for them to be a module but also to
make sure the states of the page loading with
checkboxes selected and that of the page having
checkboxes selected afterwards are both
represented when needed and in the right place in
the test.

Added bonus: removes the use of jQuery as well.