Test that user with manage api key permission can see some settings views
Even when they don't have "manage settings" permission.

While looking around, I also found that this type of user will
see all settings page rows that a user with "manage settings" would
see, but for most of them, when clicking, they would see
"you are not allowed to see this page".

This is not a good experience - we don't tell users which views
they can access. We should probably fix this and only show
them the views that they can access.