Fix occasionally misaligned placeholder highlighting
The hidden copy of the template content we use to render the yellow
placeholders was slightly wider than the original text.

This sometimes mean the text didn’t break onto a new line in the same
place. This would cause the yellow highlight to not appear in the same
place as the double brackets.

This was caused by the padding and border of the hidden copy being added
to it’s width, not subtracted from its width. `box-sizing` is the CSS
property which controls whether padding and borders are added to or
subtracted from an element’s width.