returned letters: handle cases with "orphaned" returned letters, showing message
given an api get_returned_letters "new-style" response (with
results wrapped in a dictionary and accompanied with an
orphaned_count), this displays a message at the top of a
returned letter listing page, explaining why the number of
letters shown in this list might not be the number that
summary pages had lead the user to expect.

should continue to work as before with "old-style" api
responses.