Add support form page to ask what the problem is
This page shows radio buttons to check if the problem is one that
would be considered an incident or not. The logic to redirect after this
page then takes a number of factors into account - the state of the
services the user has, whether the issue would be considered an
incident, and whether the user is logged in.