Add support form page to ask what the problem is
This page shows radio buttons to check if the problem is one that
would be considered an incident or not. The user is always redirected
to thge page with the feedback form after this, but with different
parameters based on the option they selected, whether they are logged
in, and whether they have live services. These parameters will get used
later to determine things such as Zendesk ticket subjects.