Add custom back links and Zendesk ticket subjects to support form
When redirecting to the support form with a ticket type of "problem" we
have been setting a value for "catgory" in the query string. This now
makes use of that unique value to set the back link for the form to the
previous page, to give different types of problems different Zendesk
subjects and to set the Notify responder for different ticket types.