Deprecate the "general" support route and ticket type
We classified tickets as "general" if they were opened by a logged out
user. We've now redesigned the support form flow so that we ask a few
more questions when users are logged out and classify these tickets as
either questions or problems instead - the "general" category is no
longer needed or used.

Visiting "/feedback/general" will now redirect to "/support" so that
users can be asked the triage questions about the issue they're seeing.