Deprecate the "triage" route and page
We no longer have a single page for triaging support responses - there
are a series of questions we need to ask. This deprecates the triage
page, and redirects users to the start of the process for reporting a
problem if the ticket needs triage.

The criteria for determining which tickets which need triage has not
changed.

One of the things we check to see if a ticket needs triage is whether
"severe" is in the query string and set to "yes" when you reach the
feedback form. Users shouldn't arrive on this page with a problem ticket
and "severe" not set, so they should only be redirected in rare cases if
this is somehow not included.