Add support pages to enter account details
There are 4 similar pages for a user to enter their account details if
they're having issues signing in. While each page contains a form for
them to fill in their details, the exact form fields and the Zendesk
ticket that get created are slightly different, so it's simpler to
have four separate routes.

Later iterations will add instructions to each type of ticket that gets
opened, so the ticket details are kept hidden from the user.