Stop signed in users seeing the signed out support form pages
Signed in users shouldn't be shown the signed out support form pages
since on the "Report a problem" page they won't see the sign in issues
radio button. However, just as an extra precaution, this change
redirects them if they do end up on any of the pages designed for signed
out users.