Extend screen mock to support element matching
The current screen mock stubs out the most
commonly used web APIs for getting the position
and/or dimension of the window or elements. But it
requires elements to exist before code we're
testing is run.

This changes it to also support mocking elements
that don't exist at that point.

This also switches the mocking to a form where all
calls for position and dimension are passed
through some logic that checks if they are mocked
and then passes mock data or falls back to the
normal implementation.