Fix empty test assertions
`called_once_with` is not a method of `unittest.Mock`.

Instead it is a mocked functions which will:
- accept any arguments
- always return something `True`-ish

This means the assertions in these tests are doing nothing, and changing
their arguements or the arguements of the methods they are mocking does
not cause the tests to fail.

This commit switches to using methods which do exist.