Add some object-level caching for sent_today count
This cache only persists through the lifetime of a request.

Normally would use `cached_property` for this but since it’s a method
we can’t.

If we don’t do this we end up calling Redis every time we look at this
property, which could result in mismatched counts if the Redis data
increments in the meantime.