Fix issue in test to trial mocking by CSS selector
One of the tests in this group checks if the JS
adjusts the first column headers when they have
different widths. The current test makes them
different by setting a width on the original. The
problem with that is that the original is cloned
to get the other one, so they will always match.

This swaps to an approach that mocks the width
communicated for the element from the clone instead,
which better matches how JS runs in a browser.

Doing this now to test the new feature where
mockPositionAndDimension can now accept a CSS
selector for the node to mock. The cloned element
doesn't exist before the JS to test is run so this
should test the functionality works.