Pre-populate Zendesk topics when creating sign in issue tickets
We pre-populate various fields on the support form when creating Zendesk
tickets, and we now want to pre-populate some of the topic fields for
the first time. The Zendesk client has been updated to allow topics to
be passed through, so we can now give the IDs of the topic fields and
the values we want them to have.