Allow unmocking of element's position/dimension
Should have been in https://github.com/alphagov/notifications-admin/pull/5612

This also rewrites existing calls to
`screenMock.reset()` to use this method instead,
so the code watching calls to position and
dimension methods/props on all elements is
retained. We instead move this to the same scope
in the test as that in which `screenMock` is
instantiated.