Rewrite FullScreenTable as an ES Module
We want to migrate all of our legacy Javascript
files to ESM, so this is the latest in the
series of conversions.

Rewrite follows the same conventions we
previously used for RadioSelect, FocusBanner,
Collapsible checkboxes, ColourPreview, FileUpload,
Autofocus, PreviewPane, CopyToClipboard and others.

Added bonus: removes the use of jQuery.

In this PR we:

- rewrite the module as ES Module
- update the tests to be ES Module and polyfill
some Javascript methods that JSDOM doesn't
return correct values
- remove old module JS file
- remove old module from the legacy bundle build

- this is module is used on the scheduling of
notifications on the "send" page and
- upload CSV emergency list preview page