Update 'custom email sender' page content
- content changes as per content doc
- add new section for dynamic selection preview
- modify radio buttons to remove dynamic preview

The intention is to have a dynamic preview that
changes based on radio selection and custom
from name input.

For prominence, the preview is moved outside of
the radios, under its own heading.

The preview hooks onto the
`use_custom_email_sender_name` form gets it's
updated content from the `service_email_sender_preview`
method.

Finally, the entire preview section is
in contained in a container that's only
visible when Javascript runs.