Add test for user sorting
A bug occurs when:
- the app is running under Python 3.13
- the list of users and invited users is not already sorted by email
  address

Depending on the alphabetical position of the invited user’s email
address, trying to sort the list of users and invited users would raise
an `AttributeError`.

***

We make users sortable by setting `__sort_attribute__` to
`"email_address"` on the `User` and `InvitedUser` objects. This lets
`JSONModel.__lt__` know how to sort any 2 objects of these classes.

Jinja uses a wrapped version of Python’s `sorted` algorithm:
https://github.com/pallets/jinja/blob/5ef70112a1ff19c05324ff889dd30405b1002044/src/jinja2/filters.py#L438

The function passed to `key` effectively transforms the list from
```python
[User(), User(), InvitedUser()]
```

To
```python
[[User()], [User()], [InvitedUser()]]
```

In versions of Python prior to 3.13 the latter list would be sorted
without doing any comparisons between the objects.

In Python 3.13 optimisations were added to use insertion sort over
binary sort in some cases. This causes comparisons to be more likely:

> [insertion sort] does significantly more compares overall on random data.

— https://github.com/python/cpython/commit/8383915031942f441f435a5ae800790116047b80

If this causes a comparision between a `User` and `InvitedUser` the call
to `InvitedUser.__eq__` fails because it starts looking for attributes
which don’t exist on the `User` object.

This commit fudges the order of the test data in such a way that it
forces this comparision, triggering the guard we added in
https://github.com/alphagov/notifications-admin/pull/5621/files