Add test for user sorting
A bug occurs when:
- the app is running under Python 3.13 (why?)
- the list of invited users is not already sorted by email address
- a user invites someone who is already a team member
- the list of users and invite users is sorted by Jinja’s `sort` filter,
  not Python’s `sorted` function

We make users sortable by setting `__sort_attribute__` to
`"email_address"` on the `User` and `InvitedUser` objects. This lets
`JSONModel.__lt__` know how to compare any 2 objects of these classes.
We also decorate `JSONModel` with `@functools.total_ordering`. This
implements other methods like `__gt__` for us automatically.

Jinja uses a wrapped version of Python’s `sorted` algorithm[1]. If the
algorithm finds 2 users with the same email address (so `__lt__` and
`__gt__` both return `False`) it does an additional check to compare the
two objects and see if they are the same (this triggers one of the
objects’ `__eq__` method).

This is to acheive the following:
> The sort is stable, it does not change the relative order of elements
> that compare equal.

– https://github.com/pallets/jinja/blob/5ef70112a1ff19c05324ff889dd30405b1002044/src/jinja2/filters.py#L408C1-L409C33

If this comparision is between a `User` and `InvitedUser` the call to
`InvitedUser.__eq__` fails because it starts looking for attributes
which don’t exist on the `User` object.

This commit fudges the test data in such a way that it forces this
comparision, triggering the guard we added in
https://github.com/alphagov/notifications-admin/pull/5621/files

***

1. https://github.com/pallets/jinja/blob/5ef70112a1ff19c05324ff889dd30405b1002044/src/jinja2/filters.py#L438