Show nudge about automation on the returned letters page
We haven’t seen great adoption of the returned letter callbacks. We
think this might be because teams who are reguarly processing returned
letters might not know about them.

We think that by adding a nudge on the page where users go to process
their returned letters that this might increase adoption.