Extend "preview email sender name" partial
Have have a partial for the preview and a basic
view is also used on the settings page,
so we're extending it for this purpose,
as the preview design here needs to match the
email preview look.
To do that we pass 'rich_preview=True" variable
to the partial include.

The 'rich' preview markup in made from
the SummrayList component.