Make all problem tickets require triage
We were currently only requiring a ticket to be triaged outside of
working hours. This changes the logic to require triage at any time of
day. Tickets are normally triaged as part of going through the support
form journey, so the `needs_triage` function changed here is only used
to catch the cases where someone has reached the feedback page
without `severe` in the query string.