Update logic that determines whether a ticket is an emergency
Before, a ticket could not be an emergency if it was in hours. The logic
has changed so "problem" tickets marked as an emergency by logged in
users are treated as a P1. But if a user is logged out, we only treat it
as an emergency out of hours (in which case they'd be redirected to the
page showing our OOH email address). We don't count tickets opened by
logged out users in hours as an emergency.