Add skeletons for slow-loading parts of dashboard
Anything which queries the notifications table from the dashboard is
slow, sometimes to the point of timing out.

Because none of the page loads until the whole page loads, this means
it takes a long time for users to get anything on screen, including the
navigation. This is an especially bad user experience because the
dashboard is the first page we take people to when they sign in.

This commit replaces those slow-loading parts of the dashboard with
skeletons, in other words form but not content. We can then use the
existing AJAX behaviour to populate those parts of the page later.

This should mean users get (most of) the page much quicker.

If the response does time out the AJAX code is clever enough not to
repeat the request. Users will see the skeleton indefinitely in this
case, but this is better than the current scenario where they just see a
`500` error page.