Rewrite dashboard tests to look at JSON response
At the moment the content of the JSON response is untested. We rely on
the whole-page response calling through to the same codepath, and
testing that.

Since we are going to modify the whole-page response to be different,
we need to ensure that what appears on the page – via the JSON response
– is still tested.