Add platform admin decorator to the set inbound number endpoint
"services/<id>/service-settings/set-inbound-number" should be a platform
only route - it's linked to from the platform admin part of the settings
table, but didn't have the right decorator. This page can be used if we
need to specify which available inbound number a service is assigned.

Receiving text message is self-service, but users access this through a
different route ("services/<id>/service-settings/receive-text-messages")
containing content about the functionality.