Use a per-resource polling interval when making AJAX requests
The polling interval is calculated dynamically based on the response
time of the resource.

On the dasjboard we now have 2 different resources on the page. If one
is fast and one is slow we don’t want the polling interval of the slow
resource to be based on the response time of the fast resource (and vice
versa).