Remove references to ‘trial’ and ‘live’ from  the initial state of the ‘Your services’ page
When a user is not a member of any orgs or
services, we no longer want to show "live" and
"trial" services headings.
Instead, we show just the "You are not a member
of any services yet", but do keep a
visually-hidden 'services heading as all of
these are in a 'nav' element. We also align the
'add' and 'join a service buttons to the left edge.

When a user is a member of an org but no services,
we show the "Organisations" heading as well as
"Services" heading with the above wording.

When a user is a member of trial, live or both
types of services, we should appropriate headings.