Rewrite AuthenticateSecurityKey as an ES Module
This PR rewrites the code into ES Module format
that we've been using.

In essence it splits the previous long chain of
.then Promise instances into separate Class
methods and replaces them with async/await pattern.

This PR doesn't tackle rewriting ErrorBanner as
an ES Module, nor does it seek to replace the CBOR
package.
There are separate tickets for those to be done
later.