Rewrite RegisterSecurityKey as an ES Module
This PR rewrites the code into ES Module format
that we've been using. Code changes closely
resemble that of AuthenticateSecurityKey,
rewritten in https://github.com/alphagov/notifications-admin/pull/5667

In essence it splits the previous long chain of
.then Promise instances into separate Class
methods and replaces them with async/await pattern.

This PR doesn't tackle rewriting ErrorBanner as
an ES Module, nor does it seek to replace the CBOR
package.
There are separate tickets for those to be done
later.