Remove markdown from tickets about sign in issues
These tickets get opened on behalf of the person who reported the issue.
We saw the markdown correctly rendered if one of our team tried opening
a ticket (as they account seemed to be trusted) but it isn't being
rendered if a "normal user" submits one of these tickets. This removes
the markdown to keep the ticket in plain text format.