Limit number of column headers displayed
The number of column headers in a CSV file is unbounded. If there are
too many it can cause so much HTML to be outputted that the page can’t
be generated in time.

This is often not a problem because `Recipients.column_headers`
de-duplicates columns, so each column would have to be unique in order
for it to be displayed.

However in some cases where there are errors we want to show the user
their spreadsheet with the columns exactly as uploaded, so it’s easier
to see what the problem is. Here we use `Recipients._raw_column_headers`
which does not de-duplicate.

This commit sets a very high upper bound on the number of columns which
will be displayed in these cases. This is hopefully higher than anyone
would ever reasonably use. Truncating like this should be enough that
they can load the page to see the error message and fix the problem.