On account creation show success banner on your-services page
Outcome of the onboarding user research,
this adds a notification banner when a user
completes their accoutn creation and are taken
to the /your-services page.

Used a H2 as this is what Design System's
Notification Banner uses - https://design-system.service.gov.uk/components/notification-banner/

There were two waus to implment this
a) add a view to app/templates/partials/flash_messages/
render it as a template to a variable and
pass that to the `flash` system

b) pass markup directly to the flash system

I opted for b). Flash system is already set up to
create the output for the type of notification
banner style we use.