Rewrite ErrorBanner as an ES Module
This PR rewrites the code into ES Module format
that we've been using.

Pages where this is used contain 3 error banners:
no-js, no webauth-api and error when autheticating/
registering.

We're using 'hidden' attribute instead of
a CSS class to hide/show banner.

Module accepts a custom selector, which we use
in authenticateSecurityKey and registerSecurityKey
modules.