Update banner markup
Instead of using 'govuk-!-display-none' class
to hide the banner, we'll use the hidden attribute
so that we don't rely on CSS to hide it.

We're adding a 'webauthn__error' class to the
banner that's used in ErrorBanner JS so we can
target it directly instead of toggling all 3.

The other two banners are for no-js and no-webauthn
support.